refactor: convert users route from modal-based to page-based CRUD
Replace modal/dialog-based CRUD operations with full-page routes to
match the tasks route pattern and improve consistency across the app.

- Create/Edit/Invite now use dedicated pages instead of nested modal routes
- Delete uses fetcher-based inline ConfirmDialog (with username confirmation)
- Extract form from Dialog wrapper into page-level UsersMutateForm component
- Move table components into _index/ subdirectory
- Flatten +shared/ into +components/ and +data/
- Remove all modal boilerplate (useState, setTimeout goBack hack)
