--- app/routes/_authenticated/users/+components/data-table-row-actions.tsx
@@ -1,53 +1,0 @@
-import { DotsHorizontalIcon } from '@radix-ui/react-icons'
-import { IconEdit, IconTrash } from '@tabler/icons-react'
-import type { Row } from '@tanstack/react-table'
-import { href, Link } from 'react-router'
-import { Button } from '~/components/ui/button'
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuSeparator,
-  DropdownMenuShortcut,
-  DropdownMenuTrigger,
-} from '~/components/ui/dropdown-menu'
-import type { User } from '../+shared/data/schema'
-
-interface DataTableRowActionsProps {
-  row: Row<User>
-}
-
-export function DataTableRowActions({ row }: DataTableRowActionsProps) {
-  return (
-    <DropdownMenu modal={false}>
-      <DropdownMenuTrigger asChild>
-        <Button
-          variant="ghost"
-          className="data-[state=open]:bg-muted flex h-8 w-8 p-0"
-        >
-          <DotsHorizontalIcon className="h-4 w-4" />
-          <span className="sr-only">Open menu</span>
-        </Button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="end" className="w-40">
-        <DropdownMenuItem asChild>
-          <Link to={href('/users/:user/update', { user: row.original.id })}>
-            Edit
-            <DropdownMenuShortcut>
-              <IconEdit size={16} />
-            </DropdownMenuShortcut>
-          </Link>
-        </DropdownMenuItem>
-        <DropdownMenuSeparator />
-        <DropdownMenuItem asChild className="text-red-500">
-          <Link to={href('/users/:user/delete', { user: row.original.id })}>
-            Delete
-            <DropdownMenuShortcut>
-              <IconTrash size={16} />
-            </DropdownMenuShortcut>
-          </Link>
-        </DropdownMenuItem>
-      </DropdownMenuContent>
-    </DropdownMenu>
-  )
-}
