+++ app/routes/_authenticated/users/_index/+components/users-columns.tsx
@@ -3,8 +3,8 @@ import LongText from '~/components/long-text'
 import { Badge } from '~/components/ui/badge'
 import { Checkbox } from '~/components/ui/checkbox'
 import { cn } from '~/lib/utils'
-import { callTypes, userTypes } from '../+shared/data/data'
-import type { User } from '../+shared/data/schema'
+import { callTypes, userTypes } from '../../+data/data'
+import type { User } from '../../+data/schema'
 import { DataTableColumnHeader } from './data-table-column-header'
 import { DataTableRowActions } from './data-table-row-actions'
