+++ app/routes/_authenticated/users/_index/+queries.server.ts
@@ -1,4 +1,4 @@
-import { users as initialUsers } from './+shared/data/users'
+import { users as initialUsers } from '../+data/users'
 
 interface ListFilteredUsersArgs {
   username: string
