--- app/routes/_authenticated/users/invite/+schema.ts
@@ -1,17 +1,0 @@
-import { coerceFormValue } from '@conform-to/zod/v4/future'
-import { z } from 'zod'
-
-export const formSchema = coerceFormValue(
-  z.object({
-    email: z.email({
-      error: (issue) =>
-        issue.input === undefined
-          ? 'Please enter your email'
-          : 'Invalid email address',
-    }),
-    role: z.enum(['superadmin', 'admin', 'manager', 'cashier'], {
-      error: 'Role is required.',
-    }),
-    desc: z.string().optional(),
-  }),
-)
